Disable keeping leaderboard.json files by default

+++ src/aocLeaderboard.ts
@@ -2,7 +2,8 @@ import { readFileSync, writeFileSync } from "fs";
 import fetch from "node-fetch";
 import { config } from "./configHelper";
 
-const DEBUG = false;
+const DUMMY_BOARD = process.env.DUMMY_BOARD === 'true';
+const KEEP_ALL_LEADERBOARDS = process.env.KEEP_ALL_LEADERBOARDS === 'true';
 
 export function getBoardUrl() {
     const link = config.leaderboardUrl;
@@ -18,7 +19,7 @@ export function getLastLeaderBoard() {
 }
 
 async function fetchLeaderBoard(): Promise<string> {
-    if (DEBUG) {
+    if (DUMMY_BOARD) {
         console.log(`Reading from file ./data/newleaderboard.json instead of fetching from [${getBoardUrl()}]`)
         return readFileSync('./data/newleaderboard.json', { encoding: 'utf8' });
     }
@@ -36,7 +37,9 @@ async function fetchLeaderBoard(): Promise<string> {
 export async function getNewLeaderBoard() {
     const newLeaderBoardJson = await fetchLeaderBoard();
     if (newLeaderBoardJson.startsWith('{') && JSON.parse(newLeaderBoardJson).members) {
-        writeFileSync(`./data/old/aocleaderboard-${Date.now()}.json`, getLastLeaderBoard());
+        if (KEEP_ALL_LEADERBOARDS) {
+            writeFileSync(`./data/old/aocleaderboard-${Date.now()}.json`, getLastLeaderBoard());
+        }
         writeFileSync('./data/aocleaderboard.json', newLeaderBoardJson);
     } else {
         throw Error('Bad leaderboard response:' + newLeaderBoardJson);
